% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccsMDxI10}
\alias{ccsMDxI10}
\title{ccsMDxI10}
\format{a \code{data.frame} with 9 variables
\describe{
  \item{dxCode}{(chr) Diagnosis code}
  \item{ccss}{(int) Identifies the sequence of line level data pertaining to a claim.}
  \item{ccssName}{(chr) Code value to lookup the patient's ICD code and description.}
  \item{dxDesc}{(chr) Indicates diagnosis was present at time of admission.}
  \item{ccsm1}{(num) Indicates diagnosis was present at time of admission.}
  \item{ccsm1Name}{(chr) Indicates diagnosis was present at time of admission.}
  \item{ccsm2}{(num) Indicates diagnosis was present at time of admission.}
  \item{ccsm2Name}{(chr) Indicates diagnosis was present at time of admission.}
  \item{dxCdTp}{(chr) Indicates diagnosis was present at time of admission.}
}}
\source{
ccsMDxI1019jun17.Rda
}
\usage{
ccsMDxI10
}
\description{
This table contains the CCS Medical Diagnosis for ICD 10 from 19jun17.
}
\keyword{datasets}
