% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccsMDxI9}
\alias{ccsMDxI9}
\title{ccsMDxI9}
\format{a \code{data.frame} with 10 variables
\describe{
  \item{dxCode}{(chr) Diagnosis code}
  \item{ccsm1}{(num) Indicates diagnosis was present at time of admission.}
  \item{ccsm1Name}{(chr) Indicates diagnosis was present at time of admission.}
  \item{ccsm2}{(num) Indicates diagnosis was present at time of admission.}
  \item{ccsm2Name}{(chr) Indicates diagnosis was present at time of admission.}
  \item{ccsm2}{(num) Indicates diagnosis was present at time of admission.}
  \item{ccsm2Name}{(chr) Indicates diagnosis was present at time of admission.}
  \item{ccsm2}{(num) Indicates diagnosis was present at time of admission.}
  \item{ccsm2Name}{(chr) Indicates diagnosis was present at time of admission.}
  \item{dxCdTp}{(chr) Indicates diagnosis was present at time of admission.}
}}
\source{
ccsMDxI1019jun17.Rda
}
\usage{
ccsMDxI9
}
\description{
This table contains the CCS Medical Diagnosis for ICD 9 from 19jun17.
}
\keyword{datasets}
