% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccsSDxI9}
\alias{ccsSDxI9}
\title{ccsSDxI9}
\format{a \code{data.frame} with 7 variables
\describe{
  \item{dxCode}{(chr) Diagnosis code}
  \item{ccs}{(num) Indicates diagnosis was present at time of admission.}
  \item{ccsCat}{(chr) Indicates diagnosis was present at time of admission.}
  \item{dxDesc}{(chr) Indicates diagnosis was present at time of admission.}
  \item{optCat}{(int) Indicates diagnosis was present at time of admission.}
  \item{optDesc}{(chr) Indicates diagnosis was present at time of admission.}
  \item{dxCdTp}{(chr) Indicates diagnosis was present at time of admission.}
}}
\source{
ccsSingle2015July17.Rda
}
\usage{
ccsSDxI9
}
\description{
This table contains the CCS Medical Diagnosis for ICD 9 from 19jun17.
}
\keyword{datasets}
