% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chordsTables.R
\docType{package}
\name{chordsTables}
\alias{chordsTables}
\alias{chordsTables-package}
\title{chordsTables: A package for generating priority tables for the CHORDS group.}
\description{
The chordsTables package contains the function, .RMD files, and data files necessary for generating all CHORDS priority tables. See the corresponding vignette to see an example of how to use this package.
}
