% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isoLang}
\alias{isoLang}
\title{isoLang}
\format{a \code{data.frame} with 5 variables
\describe{
  \item{code3B}{(chr) Diagnosis code}
  \item{code3T}{(chr) Indicates diagnosis was present at time of admission.}
  \item{cod2}{(chr) Indicates diagnosis was present at time of admission.}
  \item{InEnglish}{(chr) Indicates diagnosis was present at time of admission.}
  \item{InFrench}{(chr) Indicates diagnosis was present at time of admission.}
}}
\source{
isoLang_10Jul2017.Rda
}
\usage{
isoLang
}
\description{
This table contains the CCS Medical Diagnosis for ICD 9 from 19jun17.
}
\keyword{datasets}
