% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prov_type}
\alias{prov_type}
\title{Provider Type}
\format{a \code{data.frame} with 2 variables
\describe{
  \item{V1}{(int) Provider category (from CHORDS_VDW_V3.1_DataModelManual)}
  \item{V2}{(chr) Provider type description.}
}}
\source{
CHORDS_VDW_V3.1_DataModelManual.doc
}
\usage{
prov_type
}
\description{
This table contains the categories and complete title for provider types from the \code{PROVIDER_SPECIALTY} table.
}
\keyword{datasets}
