% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specialties}
\alias{specialties}
\title{Provider Specialties}
\format{a \code{data.frame} with 2 variables
\describe{
  \item{SPECIALTY}{(int) Provider specialty category (from CHORDS_VDW_V3.1_DataModelManual)}
  \item{Description}{(chr) Provider specialty description.}
}}
\source{
CHORDS_VDW_V3.1_DataModelManual.doc
}
\usage{
specialties
}
\description{
This table contains the categories and complete title for provider specialties from the \code{PROVIDER_SPECIALTY} table.
}
\keyword{datasets}
