% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateCnty}
\alias{stateCnty}
\title{State and County Fips data}
\format{a \code{data.frame} with 7 variables
\describe{
  \item{state}{(chr) Diagnosis code}
  \item{stateFP}{(chr) Indicates diagnosis was present at time of admission.}
  \item{countyFP}{(chr) Indicates diagnosis was present at time of admission.}
  \item{countyName}{(chr) Indicates diagnosis was present at time of admission.}
  \item{ClassFP}{(chr) Indicates diagnosis was present at time of admission.}
  \item{stateCnty}{(chr) Indicates diagnosis was present at time of admission.}
  \item{stCntyNm}{(chr) Indicates diagnosis was present at time of admission.}
}}
\source{
StateCountyFips15jun17.Rda
}
\usage{
stateCnty
}
\description{
This table contains the CCS Medical Diagnosis for ICD 9 from 19jun17.
}
\keyword{datasets}
