% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valSets}
\alias{valSets}
\title{Value Sets}
\format{a \code{data.frame} with 4 variables
\describe{
  \item{tableName}{(chr) Diagnosis code}
  \item{columnName}{(chr) Indicates diagnosis was present at time of admission.}
  \item{code}{(chr) Indicates diagnosis was present at time of admission.}
  \item{decode}{(chr) Indicates diagnosis was present at time of admission.}
}}
\source{
StateCountyFips15jun17.Rda
}
\usage{
valSets
}
\description{
This table contains the CCS Medical Diagnosis for ICD 9 from 19jun17.
}
\keyword{datasets}
