% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Run code to generate priority tables.}
\usage{
run_report(priority, dbserver = "NONE", dbname = "NONE",
  dbuser = "NONE", dbpassword = "NONE", outputdir = NULL,
  batchmode = "FALSE")
}
\arguments{
\item{priority}{A character vector.  This is the priority table for which you want to generate the report. Options are "P1", "P2", or "P3".}

\item{dbserver}{A string vector.  The name of the database server name. This is selected via an html window.}

\item{dbname}{A string vector.  The name of the database.  This is selected via an html window.}

\item{dbuser}{A string vector. The username that is used to connect to the database. Should be an empty string for windows authentication}

\item{dbpassword}{A string vector.  The password used to connect to the database.  Should be an empty string for windows authentication}

\item{outputdir}{A string vector.  The directory for the output document.  If none is specified, default is \code{C:/Users/<username>/Documents}}

\item{batchmode}{A string vector.  Parameter to help determine if the code is being run via batch mode or requires the UI for input parameters}
}
\value{
Creates a word document generated from an .Rmd file. The file is located in \code{C:/Users/<username>/Documents} folder (the My Documents folder for the user who generated the report) and the name of the file is the value of the priority argument (e.g., P1.docx)
}
\description{
Run code to generate priority tables.
}
\examples{
\dontrun{
run_report("P1")
}
}
